Test all the extra pins on a serial connection.
Use this program with an external testor to test all the extra pins of a serial connection.

Some of the cheap chinese USB to Serial converters don't actually controll all of the serial pins.
Use this program with something like a multimeter to test if each pin is really working or not.